# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Metrics storage service",
        "title": "MetricsStore API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get metrics list",
                "operationId": "getMetric",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Check storage status",
                "operationId": "pingMetric",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update metric with json format",
                "operationId": "updateJSONMetric",
                "parameters": [
                    {
                        "description": "Metric's data",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_arefev_mtrcstore_internal_server_model.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metric's data",
                        "schema": {
                            "$ref": "#/definitions/github_com_arefev_mtrcstore_internal_server_model.Metric"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update/{type}/{name}/{value}": {
            "post": {
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update metric by type and name",
                "operationId": "updateMetric",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metric type [counter, gauge]",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "metric value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/updates/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Mass update metrics with json format",
                "operationId": "updatesMetric",
                "parameters": [
                    {
                        "description": "Metric's data",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_arefev_mtrcstore_internal_server_model.Metric"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/value/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get metric info with json format",
                "operationId": "findJSONMetric",
                "parameters": [
                    {
                        "description": "Metric's data",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_arefev_mtrcstore_internal_server_model.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metric's data",
                        "schema": {
                            "$ref": "#/definitions/github_com_arefev_mtrcstore_internal_server_model.Metric"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/value/{type}/{name}": {
            "get": {
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Find metric by type and name",
                "operationId": "findMetric",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metric type [counter, gauge]",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "metric's value, for example 200.4",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_arefev_mtrcstore_internal_server_model.Metric": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "metric value in case of counter transfer",
                    "type": "integer"
                },
                "id": {
                    "description": "metric name",
                    "type": "string"
                },
                "type": {
                    "description": "parameter that takes the value gauge or counter",
                    "type": "string"
                },
                "value": {
                    "description": "metric value in case of gauge transfer",
                    "type": "number"
                }
            }
        }
    },
    "tags": [
        {
            "description": "\"Group of requests to get metrics\"",
            "name": "Info"
        },
        {
            "description": "\"Group of requests to update metrics\"",
            "name": "Update"
        }
    ]
}